==> 02_eICU/03_eicu_lab.ipynb <==
# Collect data from lab

import gc
import gzip
import pandas as pd

# Diagnosis file 
admissions_path = r'CSV\Exports\o02_eicu_unique_admissions.csv'

compressed_file_path = r"..\00_Datasets\eicu-2_0\lab.csv.gz"

df = pd.read_csv(compressed_file_path, compression='gzip')

# Read the specific ICD codes from a CSV file
admissions_df = pd.read_csv(admissions_path)

merged_df = admissions_df.merge(df, on='patientunitstayid', how='inner')

# List of columns to be dropped
columns_to_drop = ['activeupondischarge', 'diagnosispriority', 'diagnosisstring', 'icd9code',
                   'diagnosispriority', 'patienthealthsystemstayid', 'diagnosispriority',
                   'hospitalid', 'wardid', 'hospitaladmitoffset', 'hospitaladmitsource',
                   'hospitaldischargeyear', 'hospitaldischargetime24', 'hospitaldischargeoffset',
                   'hospitaldischargestatus', 'unittype', 'unitadmitsource', 'unitstaytype',
                   'unitdischargetime24', 'unitdischargeoffset', 'unitdischargelocation',
                   'order_of_appearance', 'labtypeid', 'labresulttext', 'labmeasurenameinterface',
                   'diagnosisoffset', 'apacheadmissiondx', 'hospitaladmittime24', 'hospitaldischargelocation',
                   'unitvisitnumber', 'labid', 'gender', 'age', 'ethnicity', 'admissionheight', 'unitadmittime24',
                   'admissionweight', 'dischargeweight', 'labresultrevisedoffset', 'unitdischargestatus', 'diagnosisid']

# Drop the specified columns
merged_df = merged_df.drop(columns=columns_to_drop)

# Filtering out rows with negative values in the labresultoffset column
merged_df = merged_df[merged_df['labresultoffset'] >= 0]

# Optional: Resetting the index if you want consecutive integer indices
merged_df.reset_index(drop=True, inplace=True)

# Export the filtered DataFrame to a CSV file
merged_df.to_csv(r'CSV\Exports\o03_eicu_lab.csv', index=False, header=True)

# Free RAM
gc.collect()
