==> 02_eICU/07_eicu_vitalaperiodic.ipynb <==
# Collecting vital aperiodic data.

import gc
import gzip
import pandas as pd

# Diagnosis file 
admissions_path = r'CSV\Exports\o02_eicu_unique_admissions.csv'

compressed_file_path = r"..\00_Datasets\eicu-2_0\vitalaperiodic.csv.gz"

df = pd.read_csv(compressed_file_path, compression='gzip')

# Read the specific ICD codes from a CSV file
admissions_df = pd.read_csv(admissions_path)

# Select only the specified columns
admissions_df = admissions_df[['uniquepid', 'patientunitstayid']]

# Merge table
merged_df = admissions_df.merge(df, on='patientunitstayid', how='inner')

# Filtering out rows with negative values in the labresultoffset column
merged_df = merged_df[merged_df['observationoffset'] >= 0]

# Replace 'vitalaperiodicid' with the actual column name you want to remove
column_to_remove = 'vitalaperiodicid'

# Use the drop method to remove the specified column
merged_df = merged_df.drop(column_to_remove, axis=1)

# Export the filtered DataFrame to a CSV file
merged_df.to_csv(r'CSV\Exports\o07_eicu_vitalaperiodic.csv', index=False, header=True)

# Free RAM
df = None
merged_df = None
admissions_df = None
gc.collect()
