==> 02_eICU/11_eicu_fill_mean_etc_table.ipynb <==
# Creating and fill the final table

import gzip
import pandas as pd
import warnings
import gc

## Creating and fill the mean table

# Read the exported for header chartevent CSV file
head_df = pd.read_csv(r'CSV\Exports\groupby\o01_eicu_grouped_mean.csv')

# Read the exported for rows chartevent CSV file
row_df = pd.read_csv(r'CSV\Exports\o09_table_rows.csv')

"""
Creating the concat column as an index
from the combination of subject_id,
hadm_id and Time_Zone.
"""

row_df["concat"] = (row_df["uniquepid"].astype(str)
                    + row_df["patientunitstayid"].astype(str)
                    + row_df["Time_Zone"].astype(str))
    
# I define the concat column as the index of the rows
row_df = row_df.set_index('concat')

"""
Creating the concat column in the table as
well which includes the measurements to be
the common point between the two tables.
"""

head_df["concat"] = (head_df["uniquepid"].astype(str)
                     + head_df["patientunitstayid"].astype(str)
                     + head_df["Time_Zone"].astype(str))

# I define the concat column as the index of the rows
head_df = head_df.set_index('concat')

# Keeping the row_df unaffected by changes.
new_df = row_df

# Suppress the specific warning.
warnings.filterwarnings("ignore", category=pd.errors.PerformanceWarning)

# Filling the table
for row in head_df.itertuples():
    con = row.Index
    item = row.observation
    value = row.Mean_Chart
    new_df.at['{}'.format(con), '{}'.format(item)] = value

# Exclude the first 3 columns and add " (Mean)" after every other header
new_df.columns = new_df.columns[:3].tolist() + [f'{col} (Mean)' for col in new_df.columns[3:]]

# Removing rows that contains no values
new_df = new_df.dropna(subset=row_df.columns[3:], how='all')

# Export the merged DataFrame to a CSV file
new_df.to_csv(r'CSV\Exports\datasets\Temp\o01_mean_table.csv', index=False)

# Free RAM
gc.collect()

## Creating and fill the median table

# Read the exported for header chartevent CSV file
head_df = pd.read_csv(r'CSV\Exports\groupby\o02_eicu_grouped_median.csv')

# Read the exported for rows chartevent CSV file
row_df = pd.read_csv(r'CSV\Exports\o09_table_rows.csv')

"""
Creating the concat column as an index
from the combination of subject_id,
hadm_id and Time_Zone.
"""

row_df["concat"] = (row_df["uniquepid"].astype(str)
                    + row_df["patientunitstayid"].astype(str)
                    + row_df["Time_Zone"].astype(str))
    
# I define the concat column as the index of the rows
row_df = row_df.set_index('concat')

"""
Creating the concat column in the table as
well which includes the measurements to be
the common point between the two tables.
"""
head_df["concat"] = (head_df["uniquepid"].astype(str)
                     + head_df["patientunitstayid"].astype(str)
                     + head_df["Time_Zone"].astype(str))

# I define the concat column as the index of the rows
head_df = head_df.set_index('concat')

# Keeping the row_df unaffected by changes.
new_df = row_df

# Suppress the specific warning.
warnings.filterwarnings("ignore", category=pd.errors.PerformanceWarning)

# Filling the table
for row in head_df.itertuples():
    con = row.Index
    item = row.observation
    value = row.Median_Chart
    new_df.at['{}'.format(con), '{}'.format(item)] = value

# Exclude the first 3 columns and add " (Mean)" after every other header
new_df.columns = new_df.columns[:3].tolist() + [f'{col} (Median)' for col in new_df.columns[3:]]

# Removing rows that contains no values
new_df = new_df.dropna(subset=row_df.columns[3:], how='all')

# Export the merged DataFrame to a CSV file
new_df.to_csv(r'CSV\Exports\datasets\Temp\o02_median_table.csv', index=False)

# Free RAM
gc.collect()

## Creating and fill the min table

# Read the exported for header chartevent CSV file
head_df = pd.read_csv(r'CSV\Exports\groupby\o03_eicu_grouped_min.csv')

# Read the exported for rows chartevent CSV file
row_df = pd.read_csv(r'CSV\Exports\o09_table_rows.csv')

"""
Creating the concat column as an index
from the combination of subject_id,
hadm_id and Time_Zone.
"""

row_df["concat"] = (row_df["uniquepid"].astype(str)
                    + row_df["patientunitstayid"].astype(str)
                    + row_df["Time_Zone"].astype(str))
    
# I define the concat column as the index of the rows
row_df = row_df.set_index('concat')

"""
Creating the concat column in the table as
well which includes the measurements to be
the common point between the two tables.
"""
head_df["concat"] = (head_df["uniquepid"].astype(str)
                     + head_df["patientunitstayid"].astype(str)
                     + head_df["Time_Zone"].astype(str))

# I define the concat column as the index of the rows
head_df = head_df.set_index('concat')

# Keeping the row_df unaffected by changes.
new_df = row_df

# Suppress the specific warning.
warnings.filterwarnings("ignore", category=pd.errors.PerformanceWarning)

# Filling the table
for row in head_df.itertuples():
    con = row.Index
    item = row.observation
    value = row.Min_Chart
    new_df.at['{}'.format(con), '{}'.format(item)] = value

# Exclude the first 3 columns and add " (Mean)" after every other header
new_df.columns = new_df.columns[:3].tolist() + [f'{col} (Min)' for col in new_df.columns[3:]]

# Removing rows that contains no values
new_df = new_df.dropna(subset=row_df.columns[3:], how='all')

# Export the merged DataFrame to a CSV file
new_df.to_csv(r'CSV\Exports\datasets\Temp\o03_min_table.csv', index=False)

# Free RAM
gc.collect()

# Creating and fill the max table

# Read the exported for header chartevent CSV file
head_df = pd.read_csv(r'CSV\Exports\groupby\o04_eicu_grouped_max.csv')

# Read the exported for rows chartevent CSV file
row_df = pd.read_csv(r'CSV\Exports\o09_table_rows.csv')

"""
Creating the concat column as an index
from the combination of subject_id,
hadm_id and Time_Zone.
"""

row_df["concat"] = (row_df["uniquepid"].astype(str)
                    + row_df["patientunitstayid"].astype(str)
                    + row_df["Time_Zone"].astype(str))
    
# I define the concat column as the index of the rows
row_df = row_df.set_index('concat')

"""
Creating the concat column in the table as
well which includes the measurements to be
the common point between the two tables.
"""
head_df["concat"] = (head_df["uniquepid"].astype(str)
                     + head_df["patientunitstayid"].astype(str)
                     + head_df["Time_Zone"].astype(str))

# I define the concat column as the index of the rows
head_df = head_df.set_index('concat')

# Keeping the row_df unaffected by changes.
new_df = row_df

# Suppress the specific warning.
warnings.filterwarnings("ignore", category=pd.errors.PerformanceWarning)

# Filling the table
for row in head_df.itertuples():
    con = row.Index
    item = row.observation
    value = row.Max_Chart
    new_df.at['{}'.format(con), '{}'.format(item)] = value

# Exclude the first 3 columns and add " (Mean)" after every other header
new_df.columns = new_df.columns[:3].tolist() + [f'{col} (Max)' for col in new_df.columns[3:]]

# Removing rows that contains no values
new_df = new_df.dropna(subset=row_df.columns[3:], how='all')

# Export the merged DataFrame to a CSV file
new_df.to_csv(r'CSV\Exports\datasets\Temp\o04_max_table.csv', index=False)

# Free RAM
gc.collect()
